requirements.txt
ArabicOcr==1.1.6
Flask==2.1.2
Pillow==9.1.1
